Rename CredentialValidationExceptionSerializer - > CredentialsAreNotValidExceptionSerializer and its test
